+++ lib/qonfig/uploaders/yaml.rb
@@ -6,13 +6,13 @@ class Qonfig::Uploaders::YAML < Qonfig::Uploaders::File
   class << self
     # @param settings [Qonfig::Settings]
     # @param options [Hash<Symbol,Any>]
-    # @option call_procs [Boolean]
+    # @param value_processor [Block]
     # @return [String]
     #
     # @api private
     # @since 0.11.0
-    def represent_settings(settings, call_procs:, **options)
-      settings_hash = settings.__to_hash__(call_procs: call_procs)
+    def represent_settings(settings, **options, &value_processor)
+      settings_hash = settings.__to_hash__(&value_processor)
       ::YAML.dump(settings_hash, **options)
     end
   end